use Double instead of Float

+++ Sources/SwiftToolbox/StringProtocol+CaseInsensitiveCompare.swift
@@ -23,20 +23,20 @@ public extension StringProtocol where Index == String.Index {
 public extension StringProtocol {
     /// Returns a value indicating the similarity of both string. If not all characters of the search match,
     /// the result is 0. If strings are same, 1 is returned
-    func match(against search: String, options mask: String.CompareOptions = [.caseInsensitive, .diacriticInsensitive, .widthInsensitive]) -> Float {
+    func match(against search: String, options mask: String.CompareOptions = [.caseInsensitive, .diacriticInsensitive, .widthInsensitive]) -> Double {
         guard !isEmpty else { return search.isEmpty ? 1 : 0 }
         guard !search.isEmpty else { return 0 }
         guard count >= search.count else { return 0 }
 
-        var similarities: Float = 0
+        var similarities: Double = 0
 
         var offset = startIndex
         for (index, letter) in search.enumerated() {
             guard let range = self[offset...].range(of: String(letter), options: mask) else { return 0 }
-            similarities += 1 / Float(1 + distance(from: startIndex, to: range.lowerBound) - index)
+            similarities += 1 / Double(1 + distance(from: startIndex, to: range.lowerBound) - index)
             offset = range.upperBound
         }
 
-        return similarities / Float(search.count)
+        return similarities / Double(search.count)
     }
 }
